Add summary of post-C&C ML discussion
This commit adds a single file summarizing the ORC mailing list
discussions which followed the Code & Compliance workshop
on 23-24 October, 2025.

The intent is to provide an accessible entrypoint for WG members
and the public who wish to understand the diversity of viewpoints
expressed after the workshop -- without combing through the
whole mailing list archive.

Signed-off-by: Aeva Black <806320+AevaOnline@users.noreply.github.com>